refactor(geo): add GeoUtils for distance calculations and update services
- Introduced GeoUtils class with CalculateDistance method using the Haversine formula.
- Updated DonorService and LocationService to utilize GeoUtils for distance calculations.
- Removed redundant distance calculation methods from both services to streamline code.
