+++ VitaFlow.Core/Common/GeoUtils.cs
@@ -1,0 +1,24 @@
+using System;
+
+namespace VitaFlow.Core.Common
+{
+    public static class GeoUtils
+    {
+        // Tính khoảng cách Haversine giữa hai điểm (km)
+        public static double CalculateDistance(double lat1, double lon1, double lat2, double lon2)
+        {
+            const double earthRadiusKm = 6371.0;
+            var dLat = ToRadians(lat2 - lat1);
+            var dLon = ToRadians(lon2 - lon1);
+            lat1 = ToRadians(lat1);
+            lat2 = ToRadians(lat2);
+            var a = Math.Sin(dLat / 2) * Math.Sin(dLat / 2) +
+                    Math.Sin(dLon / 2) * Math.Sin(dLon / 2) *
+                    Math.Cos(lat1) * Math.Cos(lat2);
+            var c = 2 * Math.Atan2(Math.Sqrt(a), Math.Sqrt(1 - a));
+            return earthRadiusKm * c;
+        }
+
+        private static double ToRadians(double angle) => angle * Math.PI / 180.0;
+    }
+}
